# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pet-microservice",
    "description": "Especificación de la API REST del microservicio desarrollado en Go relativo al manejo de información de mascotas.\n\nEnlace al repositorio:\n- [Pet-microservice repository](https://github.com/JaimePalomo/pet-microservice)",
    "contact": {
      "email": "palomosivianesjaime@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://3.124.185.228:80"
    }
  ],
...
  "paths": {
    "/creamascota": {
      "post": {
        "tags": [
          "Pets"
        ],
        "summary": "Crea una nueva mascota",
        "description": "Crea una nueva mascota y la añade a la base de datos con todos sus atributos. Todos los parámetros son necesarios. Edad y fecha de nacimiento de la mascota deben ser coherentes. Fecha de nacimiento debe tener el formato 01/02/2006. El género debe ser M o H.",
        "operationId": "createPet",
        "requestBody": {
          "description": "Crea una nueva mascota",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Mascota creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestCreaMascota"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Pet": {
        "required": [
          "nombre",
          "especie",
          "genero",
          "edad",
          "fecha_nacimiento"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "tobi"
          },
          "especie": {
            "type": "string",
            "example": "perro"
          },
          "genero": {
            "type": "string",
            "example": "M"
          },
          "edad": {
            "type": "integer",
            "example": 10
          },
          "fecha_nacimiento": {
            "type": "string",
            "example": "01/01/2012"
          }
        }
      },
...
      "BadRequestCreaMascota": {
        "type": "string",
        "example": "invalid json body"
      },
...
    }
  }
}